test: extract KeychainTestSupport (Keychain + LaunchAtLogin mocks) from TestSupport (#555, #434)
Fifth slice of splitting the large TestSupport.swift into domain files.
This one is security-relevant: MockKeychainService is the shared fake
that every credential-adjacent test suite uses (SettingsStoreTests,
KeychainSecretStoreTests, SecretSlotTests), and its addressing scheme —
a [String: String] dict keyed "service::account" — IS the storage
contract in tests. Any field/method/key-derivation drift would silently
break every Keychain round-trip test.

Moves MockKeychainService (KeychainServicing) and MockLaunchAtLoginService
(LaunchAtLoginControlling) byte-identical from TestSupport.swift into a
new KeychainTestSupport.swift with the minimum imports (Foundation plus
@testable import BugNarrator).

Pre-build codex spec-review provided byte-exact signature contract; all
11 grep checks pass on the new file:
- final class MockKeychainService: KeychainServicing {
- var values: [String: String] = [:]
- var setError: Error?, var deleteError: Error?
- var interactionRequiredKeys: Set<String> = []
- private(set) var readRequests: [(service: String, account: String, allowInteraction: Bool)] = []
- func string/setString/deleteValue signatures
- private func key(forService:account:) helper (private METHOD, not
  fileprivate — moves with the class)
- exact key-derivation body "\(service)::\(account)"

No credential-material leakage (no print/NSLog/Logger/os_log in the
mock — verified by codex). Neither mock has isolation annotations. The
KeychainServicing default extension in ServiceProtocols.swift remains
reachable via @testable.

Every other mock and AppStateHarness stay in TestSupport.swift. The full
BugNarratorTests suite passes at 662 tests (unchanged from main);
docs-sync validation passes.

Pre-build spec reviewed by codex in a fresh session (5 findings folded
in: byte-exact signature contract, grep-verifiable acceptance,
protocol-extension reachability, no-logging-regression guard, current
337-395 range not stale line numbers).
